requirements.txt
agno
google-genai
python-dotenv
yfinance
duckduckgo-search
ddgs
google-generativeai